исправляет селект
